Revert "Update shaders"
This reverts commit cb8b9ef203c82254cc0db78c2d8aeeafb50c8c85.

+++ res/shaders/basics/basic-triangle/basic-triangle.frag
@@ -1,6 +1,6 @@
 #version 300 es
-layout(location = 0) out highp vec4 color;
+layout(location = 0) out lowp vec4 color;
 
 void main() {
-	color = vec4(1.0f, 0.0f, 0.0f, 1.0f);
+	color = vec4(1, 0, 0, 1);
 }
